refactor: extract runtime offload policy
